Changed up serial code to be more in line with a working product.
