--- serial_code.c
@@ -1,26 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <omp.h>
-
-int main(int argc, int **argv)
-{
-    int num_updates = 100;
-
-
-    int track_size = 20;
-
-    int num_trains = 5;
-    int num_stations = 4;
-
-    for(int i = 0; i < num_trains ; ++i)
-    {
-        Train t = new Train();
-    }
-
-
-    for(int i = 0; i < num_updates; ++i)
-    {
-
-
-    }
-}
